commands improvement

+++ src/test/java/org/walkmod/WalkmodDispatcherTest.java
@@ -138,7 +138,11 @@ public class WalkmodDispatcherTest {
 			String userDir = new File(System.getProperty("user.dir")).getAbsolutePath();
 			System.setProperty("user.dir", tmp.getAbsolutePath());
 			try {
-				run(new String[] { "add-plugin", "--groupId", "org.walkmod", "--artifactId", "walkmod-imports-cleaner-plugin", "--version", "2.0" });
+				run(new String[] { "add-plugin",  "org.walkmod:walkmod-imports-cleaner-plugin:2.0" });
+			}
+			catch(Exception e){
+				e.printStackTrace();
+				Assert.assertFalse(false);
 			}
 			finally {
 				System.setProperty("user.dir", userDir);
@@ -151,6 +155,39 @@ public class WalkmodDispatcherTest {
 				System.out.println(content);
 				
 				Assert.assertTrue(content.contains("walkmod-imports-cleaner-plugin"));
+				
+				
+				FileUtils.deleteDirectory(tmp);
+			}
+			
+		}
+	}
+	
+	@Test
+	public void testTranformationPlugin() throws Exception{
+		File tmp = new File("src/test/resources/initTestAddTrans");
+		tmp.mkdirs();
+		if(tmp.exists()){
+			String userDir = new File(System.getProperty("user.dir")).getAbsolutePath();
+			System.setProperty("user.dir", tmp.getAbsolutePath());
+			try {
+				run(new String[] { "add",  "imports-cleaner" });
+			}
+			catch(Exception e){
+				e.printStackTrace();
+				Assert.assertFalse(false);
+			}
+			finally {
+				System.setProperty("user.dir", userDir);
+				
+				File cfg = new File(tmp, "walkmod.xml");
+				Assert.assertTrue(cfg.exists());
+				
+				String content = FileUtils.readFileToString(cfg);
+				
+				System.out.println(content);
+				
+				Assert.assertTrue(content.contains("imports-cleaner"));
 				
 				
 				FileUtils.deleteDirectory(tmp);
